finished the relations and url routes with get, put, post and delete requests, prepared backend useable for frontend

+++ comments/views.py
@@ -9,11 +9,14 @@ from emojis.models import Emoji
 from .models import Comment
 
 class Comments(APIView):
-  # test function to get all comments
-  def get(self, request):
+  
+  # get all comments from single emoji
+  def get(self, request, emojiId):
+    emoji = Emoji.objects.get(pk=emojiId)
     comments = Comment.objects.all()
-    serializer = CommentSerializer(comments, many=True)
-
+    send = comments.filter(emoji=emoji.id)
+    
+    serializer = CommentSerializer(send, many=True)
     return Response(serializer.data, status=status.HTTP_200_OK)
   
   def post(self, request, emojiId):
@@ -34,11 +37,33 @@ class CommentDetail(APIView):
     except Comment.DoesNotExist:
       raise NotFound()
 
-  def is_comment_user(self, comment , user):
+  def is_comment_user(self, comment, user):
         if comment.user.id != user.id:
             raise PermissionDenied()
-  def delete(self, request, pk):
+
+  def get(self, request, pk, emojiId):
+    comment = self.get_one(pk=pk)
+    if str(comment.emoji.id) != str(emojiId):
+        return Response({'msg': 'Emoji not found.'}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+    serializer = CommentSerializer(comment)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+  def put(self, request, pk, emojiId):
+    request.data['user'] = request.user.id
+    comment = self.get_one(pk=pk)
+    if str(comment.emoji.id) != str(emojiId):
+        return Response({'msg': 'Emoji not found.'}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+    serializer = CommentSerializer(comment, data=request.data)
+    if serializer.is_valid():
+      serializer.save()
+      return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response(serializer.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+    
+  def delete(self, request, pk, emojiId):
+      request.data['user'] = request.user.id
       comment = self.get_one(pk=pk)
+      if str(comment.emoji.id) != str(emojiId):
+        return Response({'msg': 'Emoji not found.'}, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
       self.is_comment_user(comment, request.user)
       comment.delete()
       return Response(status=status.HTTP_204_NO_CONTENT)
